Language.LambdaPi.Impl.FreeFoil

Description

Free foil implementation of the \(\lambda\Pi\)-calculus (with pairs).

Free foil provides general definitions or implementations for the following:

Freely generated (from a simple signature) scope-safe AST.

Correct capture-avoiding substitution (see substitute).

Correct \(\alpha\)-equivalence checks (see alphaEquiv and alphaEquivRefreshed) as well as \(\alpha\)-normalization (see refreshAST).

Conversion helpers (see convertToAST and convertFromAST).

The following is implemented manually in this module:

Convenient pattern synonyms.

ZipMatch instances (enabling general \(\alpha\)-equivalence).

Conversion between scope-safe and raw term representation (the latter is generated via BNFC), see toLambdaPi and fromLambdaPi.

Computation of weak head normal form (WHNF), see whnf.

Entry point, gluing everything together. See defaultMain.

Note: free foil does not (easily) support patterns at the moment, so only wildcard patterns and variable patterns are handled in this implementation.

See Language.LambdaPi.Impl.FreeFoilTH for a variation of this with more automation via Template Haskell.

Documentation

>>> import qualified Control.Monad.Foil as Foil
>>> import Control.Monad.Free.Foil
>>> :set -XOverloadedStrings
>>> :set -XDataKinds

data LambdaPiF scope term

The signature Bifunctor for the \(\lambda\Pi\).

Constructors

AppF term term
Application: \((t_1 \; t_2)\)

LamF scope
Abstraction: \(\lambda x. t\)

PiF term scope
Dependent function type: \(\prod_{x : T_1} T_2\)

UniverseF
Universe (type of types): \(\mathcal{U}\)

data PairF scope term

The signature Bifunctor for pairs.

Constructors

PairF term term
Pair: \(\langle t_1, t_2 \rangle\)

FirstF term
First projection: \(\pi_1(t)\)

SecondF term
Second projection: \(\pi_2(t)\)

ProductF term term
Product type (non-dependent): \(T_1 \times T_2\)

type (:+:) = Sum :: (Type -> Type -> Type) -> (Type -> Type -> Type) -> Type -> Type -> Type

Sum of signature bifunctors.

type LambdaPi (n :: S) = AST NameBinder (LambdaPiF :+: PairF) n

\(\lambda\Pi\)-terms in scope n, freely generated from the sum of signatures LambdaPiF and PairF.

pattern App :: LambdaPi n -> LambdaPi n -> LambdaPi n

pattern Lam :: () => NameBinder n l -> LambdaPi l -> LambdaPi n

pattern Pi :: () => NameBinder n l -> LambdaPi n -> LambdaPi l -> LambdaPi n

pattern Pair :: LambdaPi n -> LambdaPi n -> LambdaPi n

pattern First :: LambdaPi n -> LambdaPi n

pattern Second :: LambdaPi n -> LambdaPi n

pattern Product :: LambdaPi n -> LambdaPi n -> LambdaPi n

pattern Universe :: LambdaPi n

whnf :: forall (n :: S). Distinct n => Scope n -> LambdaPi n -> LambdaPi n

Compute weak head normal form (WHNF) of a \(\lambda\Pi\)-term.

>>> whnf Foil.emptyScope "(λx.(λ_.x)(λy.x))(λy.λz.z)"
λ x0 . λ x1 . x1

>>> whnf Foil.emptyScope "(λs.λz.s(s(z)))(λs.λz.s(s(z)))"
λ x1 . (λ x0 . λ x1 . x0 (x0 x1)) ((λ x0 . λ x1 . x0 (x0 x1)) x1)

Note that during computation bound variables can become unordered in the sense that binders may easily repeat or decrease. For example, in the following expression, inner binder has lower index that the outer one:

>>> whnf Foil.emptyScope "(λx.λy.x)(λx.x)"
λ x1 . λ x0 . x0

At the same time, without substitution, we get regular, increasing binder indices:

>>> "λx.λy.y" :: LambdaPi Foil.VoidS
λ x0 . λ x1 . x1

To compare terms for \(\alpha\)-equivalence, we may use alphaEquiv:

>>> alphaEquiv Foil.emptyScope (whnf Foil.emptyScope "(λx.λy.x)(λx.x)") "λx.λy.y"
True

We may also normalize binders using refreshAST:

>>> refreshAST Foil.emptyScope (whnf Foil.emptyScope "(λx.λy.x)(λx.x)")
λ x0 . λ x1 . x1

toLambdaPiLam

Arguments

:: forall (n :: S). Distinct n

=> Scope n
Target scope of the \(\lambda\Pi\)-term.

-> Map VarIdent (Name n)
Mapping for the free variables in the raw term.

-> Pattern
Raw pattern (argument) of the \(\lambda\)-abstraction.

-> ScopedTerm
Raw body of the \(\lambda\)-abstraction.

-> LambdaPi n

Convert a raw \(\lambda\)-abstraction into a scope-safe \(\lambda\Pi\)-term.

toLambdaPiPi

Arguments

:: forall (n :: S). Distinct n

=> Scope n
Target scope of the \(\lambda\Pi\)-term.

-> Map VarIdent (Name n)
Mapping for the free variables in the raw term.

-> Pattern
Raw argument pattern of the \(\Pi\)-type.

-> Term
Raw argument type of the \(\Pi\)-type.

-> ScopedTerm
Raw body (result type family) of the \(\Pi\)-type.

-> LambdaPi n

Convert a raw \(\Pi\)-type into a scope-safe \(\lambda\Pi\)-term.

toLambdaPi

Arguments

:: forall (n :: S). Distinct n

=> Scope n
Target scope of the \(\lambda\Pi\)-term.

-> Map VarIdent (Name n)
Mapping for the free variables in the raw term.

-> Term
Raw expression or type.

-> LambdaPi n

Convert a raw expression into a scope-safe \(\lambda\Pi\)-term.

toLambdaPiClosed :: Term -> LambdaPi 'VoidS

Convert a raw expression into a closed scope-safe \(\lambda\Pi\)-term.

fromLambdaPi

Arguments

:: forall (n :: S). [VarIdent]
Stream of fresh raw identifiers.

-> NameMap n VarIdent
A total map for all names in scope n.

-> LambdaPi n
A scope-safe \(\lambda\Pi\)-term.

-> Term

Convert back from a scope-safe \(\lambda\Pi\)-term into a raw expression or type.

fromLambdaPi'

Arguments

:: forall (n :: S). LambdaPi n
A scope-safe \(\lambda\Pi\)-term.

-> Term

Convert back from a scope-safe \(\lambda\Pi\)-term into a raw expression or type.

In contrast to fromLambdaPi, this function uses the raw foil identifiers (integers) to generate names for the variables. This makes them transparent when printing.

ppLambdaPi :: LambdaPi 'VoidS -> String

Pretty-print a closed \(\lambda\Pi\)-term.

interpretCommand :: Command -> IO ()

Interpret a \(\lambda\Pi\) command.

interpretProgram :: Program -> IO ()

Interpret a \(\lambda\Pi\) program.

defaultMain :: IO ()

A \(\lambda\Pi\) interpreter implemented via the free foil.