Language.LambdaPi.Impl.FreeFoilTH

Description

Free foil implementation of the \(\lambda\Pi\)-calculus (with pairs).

Free foil provides general definitions or implementations for the following:

Freely generated (from a simple signature) scope-safe AST.

Correct capture-avoiding substitution (see substitute).

Correct \(\alpha\)-equivalence checks (see alphaEquiv and alphaEquivRefreshed) as well as \(\alpha\)-normalization (see refreshAST).

Conversion helpers (see convertToAST and convertFromAST).

The following is provided via generic representation via kind-generics (see Generics.Kind): 1. ZipMatch instances for signatures (enabling general \(\alpha\)-equivalence). 2. Sinkable instances for terms. 3. CoSinkable instances for patterns.

The following is generated using Template Haskell:

Signature bifunctor for terms.

Convenient pattern synonyms.

Conversion between scope-safe and raw term representation.

The following is implemented manually in this module:

Computation of weak head normal form (WHNF), see whnf.

Entry point, gluing everything together. See defaultMain.

Note: free foil does not (easily) support patterns at the moment, so only wildcard patterns and variable patterns are handled in this implementation.

Documentation

>>> :set -XOverloadedStrings
>>> :set -XDataKinds
>>> import qualified Control.Monad.Foil as Foil
>>> import Control.Monad.Free.Foil
>>> import Data.String (fromString)

Generated code

Signature

data Term'Sig a scope term

Generated with mkSignature. A signature bifunctor, specifying the nodes of a syntax tree corresponding to Term'.

Constructors

PiSig a term scope
Corresponds to Pi.

LamSig a scope
Corresponds to Lam.

AppSig a term term
Corresponds to App.

ProductSig a term term
Corresponds to Product.

PairSig a term term
Corresponds to Pair.

FirstSig a term
Corresponds to First.

SecondSig a term
Corresponds to Second.

UniverseSig a
Corresponds to Universe.

Pattern synonyms

pattern Universe :: a -> AST binder (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type UniverseSig.

pattern Second :: a -> AST binder (Term'Sig a) n -> AST binder (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type SecondSig.

pattern First :: a -> AST binder (Term'Sig a) n -> AST binder (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type FirstSig.

pattern Pair :: a -> AST binder (Term'Sig a) n -> AST binder (Term'Sig a) n -> AST binder (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type PairSig.

pattern Product :: a -> AST binder (Term'Sig a) n -> AST binder (Term'Sig a) n -> AST binder (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type ProductSig.

pattern App :: a -> AST binder (Term'Sig a) n -> AST binder (Term'Sig a) n -> AST binder (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type AppSig.

pattern Lam :: () => a -> binder n l1 -> AST binder (Term'Sig a) l1 -> AST binder (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type LamSig.

pattern Pi :: () => a -> AST binder (Term'Sig a) n -> binder n l2 -> AST binder (Term'Sig a) l2 -> AST binder (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type PiSig.

Conversion helpers

convertToTerm'Sig :: Term' a -> Either VarIdent (Term'Sig a (Pattern' a, ScopedTerm' a) (Term' a))

Generated with mkConvertToSig. Perform one step of converting Term', peeling off one node of type Term'Sig.

getPattern'Binder :: Pattern' a -> Maybe VarIdent

Generated with mkGetPatternBinder. Extract at most one binder from a pattern or crash.

getTerm'FromScopedTerm' :: ScopedTerm' a -> Term' a

Generated with mkGetScopedTerm. Extract scoped term or crash.

convertFromTerm'Sig :: Term'Sig a (Pattern' a, ScopedTerm' a) (Term' a) -> Term' a

Generated with mkConvertFromSig. Perform one step of converting Term', peeling off one node of type Term'Sig.

Scope-safe patterns

data FoilPattern' a (n :: S) (l :: S) where

Generated with mkFoilPattern. A scope-safe version of Pattern'.

Constructors

FoilPatternWildcard :: forall a (n :: S). a -> FoilPattern' a n n
Corresponds to PatternWildcard.

FoilPatternVar :: forall a (n :: S) (l :: S). a -> NameBinder n l -> FoilPattern' a n l
Corresponds to PatternVar.

FoilPatternPair :: forall a (n :: S) (n2 :: S) (l :: S). a -> FoilPattern' a n n2 -> FoilPattern' a n2 l -> FoilPattern' a n l
Corresponds to PatternPair.

toFoilPattern' :: forall (n :: S) a r. Distinct n => Scope n -> Map VarIdent (Name n) -> Pattern' a -> (forall (l :: S). DExt n l => FoilPattern' a n l -> Map VarIdent (Name l) -> r) -> r

Generated with mkToFoilPattern. Convert a raw pattern into a scope-safe pattern.

fromFoilPattern' :: forall a (n :: S) (l :: S). (Int -> VarIdent) -> FoilPattern' a n l -> Pattern' a

Generated with mkFromFoilPattern. Convert a scope-safe pattern into a raw pattern.

User-defined code

type Term' a = AST (FoilPattern' a) (Term'Sig a)

Generic annotated scope-safe \(\lambda\Pi\)-terms with patterns.

type Term = Term' BNFC'Position

Scode-safe \(\lambda\Pi\)-terms annotated with source code position.

type FoilPattern = FoilPattern' BNFC'Position

Scope-safe patterns annotated with source code position.

Conversion helpers

toTerm' :: forall (n :: S) a. Distinct n => Scope n -> Map VarIdent (Name n) -> Term' a -> Term' a n

Convert Term' into a scope-safe term. This is a special case of convertToAST.

toTerm'Closed :: Term' a -> Term' a 'VoidS

Convert Term' into a closed scope-safe term. This is a special case of toTerm'.

fromTerm' :: forall a (n :: S). Term' a n -> Term' a

Convert a scope-safe representation back into Term'. This is a special case of convertFromAST.

VarIdent names are generated based on the raw identifiers in the underlying foil representation.

This function does not recover location information for variables, patterns, or scoped terms.

Evaluation

matchPattern :: forall (n :: S) (l :: S). FoilPattern n l -> Term n -> Substitution Term l n

Match a pattern against an term.

whnf :: forall (n :: S). Distinct n => Scope n -> Term n -> Term n

Compute weak head normal form (WHNF) of a \(\lambda\Pi\)-term.

>>> whnf Foil.emptyScope "(λx.(λ_.x)(λy.x))(λ(y,z).z)"
λ (x0, x1) . x1

>>> whnf Foil.emptyScope "(λs.λz.s(s(z)))(λs.λz.s(s(z)))"
λ x1 . (λ x0 . λ x1 . x0 (x0 x1)) ((λ x0 . λ x1 . x0 (x0 x1)) x1)

Note that during computation bound variables can become unordered in the sense that binders may easily repeat or decrease. For example, in the following expression, inner binder has lower index that the outer one:

>>> whnf Foil.emptyScope "(λx.λy.x)(λx.x)"
λ x1 . λ x0 . x0

At the same time, without substitution, we get regular, increasing binder indices:

>>> "λx.λy.y" :: Term Foil.VoidS
λ x0 . λ x1 . x1

To compare terms for \(\alpha\)-equivalence, we may use alphaEquiv:

>>> alphaEquiv Foil.emptyScope (whnf Foil.emptyScope "(λx.λy.x)(λx.x)") "λx.λy.y"
True

We may also normalize binders using refreshAST:

>>> refreshAST Foil.emptyScope (whnf Foil.emptyScope "(λx.λy.x)(λx.x)")
λ x0 . λ x1 . x1

\(\lambda\Pi\)-interpreter

interpretCommand :: Command -> IO ()

Interpret a \(\lambda\Pi\) command.

interpretProgram :: Program -> IO ()

Interpret a \(\lambda\Pi\) program.

defaultMain :: IO ()

A \(\lambda\Pi\) interpreter implemented via the free foil.

Orphan instances

UnifiableInPattern BNFC'Position
Ignoring location information when unifying patterns.

Instance details

Methods

unifyInPattern :: BNFC'Position -> BNFC'Position -> Bool

ZipMatchK BNFC'Position
Ignore BNFC'Position when matching terms.

Instance details

Methods

zipMatchWithK :: forall (as :: LoT Type) (bs :: LoT Type) (cs :: LoT Type). Mappings as bs cs -> (BNFC'Position :@@: as) -> (BNFC'Position :@@: bs) -> Maybe (BNFC'Position :@@: cs)