Language.LambdaPi.Impl.FreeFoilTH

Description

Free foil implementation of the \(\lambda\Pi\)-calculus (with pairs).

Free foil provides general definitions or implementations for the following:

Freely generated (from a simple signature) scope-safe AST.

Correct capture-avoiding substitution (see substitute).

Correct α-equivalence checks (see alphaEquiv and alphaEquivRefreshed) as well as α-normalization (see refreshAST).

Conversion helpers (see convertToAST and convertFromAST).

The following is generated using Template Haskell:

Convenient pattern synonyms.

ZipMatch instances (enabling general α-equivalence).

Conversion between scope-safe and raw term representation.

The following is implemented manually in this module:

Computation of weak head normal form (WHNF), see whnf.

Entry point, gluing everything together. See defaultMain.

Note: free foil does not (easily) support patterns at the moment, so only wildcard patterns and variable patterns are handled in this implementation.

Documentation

>>> :set -XOverloadedStrings
>>> :set -XDataKinds
>>> import qualified Control.Monad.Foil as Foil
>>> import Control.Monad.Free.Foil
>>> import Data.String (fromString)

Generated code

Signature

data Term'Sig a scope term

Generated with mkSignature. A signature bifunctor, specifying the nodes of a syntax tree corresponding to Term'.

Constructors

PiSig a term scope
Corresponds to Pi.

LamSig a scope
Corresponds to Lam.

AppSig a term term
Corresponds to App.

ProductSig a term term
Corresponds to Product.

PairSig a term term
Corresponds to Pair.

FirstSig a term
Corresponds to First.

SecondSig a term
Corresponds to Second.

UniverseSig a
Corresponds to Universe.

Pattern synonyms

pattern Universe :: a -> AST (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type UniverseSig.

pattern Second :: a -> AST (Term'Sig a) n -> AST (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type SecondSig.

pattern First :: a -> AST (Term'Sig a) n -> AST (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type FirstSig.

pattern Pair :: a -> AST (Term'Sig a) n -> AST (Term'Sig a) n -> AST (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type PairSig.

pattern Product :: a -> AST (Term'Sig a) n -> AST (Term'Sig a) n -> AST (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type ProductSig.

pattern App :: a -> AST (Term'Sig a) n -> AST (Term'Sig a) n -> AST (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type AppSig.

pattern Lam :: () => a -> NameBinder n l1 -> AST (Term'Sig a) l1 -> AST (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type LamSig.

pattern Pi :: () => a -> AST (Term'Sig a) n -> NameBinder n l2 -> AST (Term'Sig a) l2 -> AST (Term'Sig a) n

Generated with mkPatternSynonyms. Pattern synonym for an AST node of type PiSig.

Conversion helpers

convertToTerm'Sig :: Term' a -> Either VarIdent (Term'Sig a (Pattern' a, ScopedTerm' a) (Term' a))

Generated with mkConvertToFreeFoil. Perform one step of converting Term', peeling off one node of type Term'Sig.

getPattern'Binder :: Pattern' a -> Maybe VarIdent

Generated with mkGetPatternBinder. Extract at most one binder from a pattern or crash.

getTerm'FromScopedTerm' :: ScopedTerm' a -> Term' a

Generated with mkGetScopedTerm. Extract scoped term or crash.

convertFromTerm'Sig :: Term'Sig a (Pattern' a, ScopedTerm' a) (Term' a) -> Term' a

Generated with mkConvertToFreeFoil. Perform one step of converting Term', peeling off one node of type Term'Sig.

User-defined code

type Term' a = AST (Term'Sig a)

type Term = Term' BNFC'Position

Conversion helpers

toTerm' :: forall (n :: S) a. Distinct n => Scope n -> Map VarIdent (Name n) -> Term' a -> AST (Term'Sig a) n

Convert Term' into a scope-safe term. This is a special case of convertToAST.

toTerm'Closed :: Term' a -> AST (Term'Sig a) 'VoidS

Convert Term' into a closed scope-safe term. This is a special case of toTerm'.

fromTerm' :: forall a (n :: S). AST (Term'Sig a) n -> Term' a

Convert a scope-safe representation back into Term'. This is a special case of convertFromAST.

VarIdent names are generated based on the raw identifiers in the underlying foil representation.

This function does not recover location information for variables, patterns, or scoped terms.

Evaluation

whnf :: forall (n :: S). Distinct n => Scope n -> Term n -> Term n

Compute weak head normal form (WHNF) of a λΠ-term.

>>> whnf Foil.emptyScope "(λx.(λ_.x)(λy.x))(λy.λz.z)"
λ x0 . λ x1 . x1

>>> whnf Foil.emptyScope "(λs.λz.s(s(z)))(λs.λz.s(s(z)))"
λ x1 . (λ x0 . λ x1 . x0 (x0 x1)) ((λ x0 . λ x1 . x0 (x0 x1)) x1)

Note that during computation bound variables can become unordered in the sense that binders may easily repeat or decrease. For example, in the following expression, inner binder has lower index that the outer one:

>>> whnf Foil.emptyScope "(λx.λy.x)(λx.x)"
λ x1 . λ x0 . x0

At the same time, without substitution, we get regular, increasing binder indices:

>>> "λx.λy.y" :: Term Foil.VoidS
λ x0 . λ x1 . x1

To compare terms for \(\alpha\)-equivalence, we may use alphaEquiv:

>>> alphaEquiv Foil.emptyScope (whnf Foil.emptyScope "(λx.λy.x)(λx.x)") "λx.λy.y"
True

We may also normalize binders using refreshAST:

>>> refreshAST Foil.emptyScope (whnf Foil.emptyScope "(λx.λy.x)(λx.x)")
λ x0 . λ x1 . x1

Type checking

λΠ-interpreter

interpretCommand :: Command -> IO ()

Interpret a λΠ command.

interpretProgram :: Program -> IO ()

Interpret a λΠ program.

defaultMain :: IO ()

A λΠ interpreter implemented via the free foil.